groups support

+++ src/database_manage.rs
@@ -23,11 +23,16 @@ async fn decrypt_string(encrypted: &[u8], keypair: &PKey<Private>) -> String {
     decrypted.truncate(decrypted_len);
     String::from_utf8(decrypted).unwrap()
 }
-pub async fn is_in_db(db_transaction: &mut Transaction<'_>, chat_id: &i64) -> Result<bool, Error> {
+pub async fn is_in_db(
+    db_transaction: &mut Transaction<'_>,
+    chat_id: &i64,
+    user_id: u64,
+) -> Result<bool, Error> {
+    let bytes = user_id.to_le_bytes().to_vec();
     Ok(db_transaction
         .execute(
-            "SELECT \"user\" , state , city FROM chat WHERE id = $1",
-            &[chat_id],
+            "SELECT \"user\" , state , city FROM chat WHERE id = $1 AND user_id = $2",
+            &[chat_id, &bytes],
         )
         .await?
         == 1)
@@ -62,136 +67,160 @@ pub async fn search_city(
 pub async fn get_client_selected(
     db_transaction: &mut Transaction<'_>,
     chat_id: &i64,
+    user_id: u64,
 ) -> Result<String, Error> {
-    let row: &Row = &search_client(db_transaction, chat_id).await?[0];
+    let row: &Row = &search_client(db_transaction, chat_id, user_id).await?[0];
     Ok(row.get("selected"))
 }
 
 pub async fn get_client_city(
     db_transaction: &mut Transaction<'_>,
     chat_id: &i64,
+    user_id: u64,
 ) -> Result<String, Error> {
-    let row: &Row = &search_client(db_transaction, chat_id).await?[0];
+    let row: &Row = &search_client(db_transaction, chat_id, user_id).await?[0];
     row.try_get("city")
 }
 
 pub async fn get_client_context(
     db_transaction: &mut Transaction<'_>,
     chat_id: &i64,
+    user_id: u64,
 ) -> Result<String, Error> {
-    let row: &Row = &search_client(db_transaction, chat_id).await?[0];
+    let row: &Row = &search_client(db_transaction, chat_id, user_id).await?[0];
     Ok(row.get("context"))
 }
 pub async fn get_client_state(
     db_transaction: &mut Transaction<'_>,
     chat_id: &i64,
+    user_id: u64,
 ) -> Result<String, Error> {
-    let row: &Row = &search_client(db_transaction, chat_id).await?[0];
+    let row: &Row = &search_client(db_transaction, chat_id, user_id).await?[0];
     Ok(row.get("state"))
 }
 
 pub async fn get_client_pattern_search(
     db_transaction: &mut Transaction<'_>,
     chat_id: &i64,
+    user_id: u64,
 ) -> Result<bool, Error> {
-    let row: &Row = &search_client(db_transaction, chat_id).await?[0];
+    let row: &Row = &search_client(db_transaction, chat_id, user_id).await?[0];
     row.try_get("pattern_search")
 }
 pub async fn search_client(
     db_transaction: &mut Transaction<'_>,
     chat_id: &i64,
+    user_id: u64,
 ) -> Result<Vec<Row>, Error> {
+    let bytes = user_id.to_le_bytes().to_vec();
     let search = db_transaction
         .prepare(
-            "SELECT \"user\" , state , city , selected , pattern_search , context FROM chat WHERE id = $1",
+            "SELECT \"user\" , state , city , selected , pattern_search , context FROM chat WHERE id = $1 AND user_id = $2",
         )
         .await?;
-    Ok(db_transaction.query(&search, &[chat_id]).await?)
+    Ok(db_transaction.query(&search, &[chat_id, &bytes]).await?)
 }
 pub async fn insert_client(
     db_transaction: &mut Transaction<'_>,
     chat_id: &i64,
+    user_id: u64,
     user: String,
     keypair: &PKey<Private>,
 ) -> Result<u64, Error> {
+    let bytes = user_id.to_le_bytes().to_vec();
     let insert = db_transaction
-        .prepare("INSERT INTO chat (id , \"user\" , state , context ) VALUES ($1 , $2 , $3 , $4 )")
+        .prepare("INSERT INTO chat (id , \"user\" , state , context , user_id ) VALUES ($1 , $2 , $3 , $4 , $5 )")
         .await?;
     let user_encrypted = encrypt_string(user, keypair).await;
     Ok(db_transaction
-        .execute(&insert, &[chat_id, &user_encrypted, &"Initial", &"Initial"])
+        .execute(
+            &insert,
+            &[chat_id, &user_encrypted, &"Initial", &"Initial", &bytes],
+        )
         .await?)
 }
 pub async fn delete_client(
     db_transaction: &mut Transaction<'_>,
     chat_id: &i64,
+    user_id: u64,
 ) -> Result<u64, Error> {
+    let bytes = user_id.to_le_bytes().to_vec();
     let delete = db_transaction
-        .prepare("DELETE FROM chat WHERE id = $1")
+        .prepare("DELETE FROM chat WHERE id = $1 AND user_id = $2")
         .await?;
-    Ok(db_transaction.execute(&delete, &[chat_id]).await?)
+    Ok(db_transaction.execute(&delete, &[chat_id, &bytes]).await?)
 }
 pub async fn modify_state(
     db_transaction: &mut Transaction<'_>,
     chat_id: &i64,
+    user_id: u64,
     new_state: String,
 ) -> Result<u64, Error> {
+    let bytes = user_id.to_le_bytes().to_vec();
     let modify_state = db_transaction
-        .prepare("UPDATE chat SET state = $1 WHERE id = $2")
+        .prepare("UPDATE chat SET state = $1 WHERE id = $2 AND user_id = $3")
         .await?;
     Ok(db_transaction
-        .execute(&modify_state, &[&new_state, chat_id])
+        .execute(&modify_state, &[&new_state, chat_id, &bytes])
         .await?)
 }
 
 pub async fn modify_context(
     db_transaction: &mut Transaction<'_>,
     chat_id: &i64,
+    user_id: u64,
     new_context: String,
 ) -> Result<u64, Error> {
+    let bytes = user_id.to_le_bytes().to_vec();
     let modify_context = db_transaction
-        .prepare("UPDATE chat SET context = $1 WHERE id = $2")
+        .prepare("UPDATE chat SET context = $1 WHERE id = $2 AND user_id = $3")
         .await?;
     Ok(db_transaction
-        .execute(&modify_context, &[&new_context, chat_id])
+        .execute(&modify_context, &[&new_context, chat_id, &bytes])
         .await?)
 }
 
 pub async fn modify_city(
     db_transaction: &mut Transaction<'_>,
     chat_id: &i64,
+    user_id: u64,
     new_city: String,
 ) -> Result<u64, Error> {
+    let bytes = user_id.to_le_bytes().to_vec();
     let modify_city = db_transaction
-        .prepare("UPDATE chat SET city = $1 WHERE id = $2")
+        .prepare("UPDATE chat SET city = $1 WHERE id = $2 AND user_id = $3")
         .await?;
     Ok(db_transaction
-        .execute(&modify_city, &[&new_city, chat_id])
+        .execute(&modify_city, &[&new_city, chat_id, &bytes])
         .await?)
 }
 pub async fn modify_selected(
     db_transaction: &mut Transaction<'_>,
     chat_id: &i64,
+    user_id: u64,
     new_selected: String,
 ) -> Result<u64, Error> {
+    let bytes = user_id.to_le_bytes().to_vec();
     let modify_selected = db_transaction
-        .prepare("UPDATE chat SET selected = $1 WHERE id = $2")
+        .prepare("UPDATE chat SET selected = $1 WHERE id = $2 AND user_id = $3")
         .await?;
     Ok(db_transaction
-        .execute(&modify_selected, &[&new_selected, chat_id])
+        .execute(&modify_selected, &[&new_selected, chat_id, &bytes])
         .await?)
 }
 
 pub async fn modify_pattern_search(
     db_transaction: &mut Transaction<'_>,
     chat_id: &i64,
+    user_id: u64,
     search_mode: bool,
 ) -> Result<u64, Error> {
+    let bytes = user_id.to_le_bytes().to_vec();
     let modify_pattern_search = db_transaction
-        .prepare("UPDATE chat SET pattern_search = $1 WHERE id = $2")
+        .prepare("UPDATE chat SET pattern_search = $1 WHERE id = $2 AND user_id = $3")
         .await?;
     Ok(db_transaction
-        .execute(&modify_pattern_search, &[&search_mode, chat_id])
+        .execute(&modify_pattern_search, &[&search_mode, chat_id, &bytes])
         .await?)
 }
 pub async fn get_city_by_pattern(
